Updated CategoryId DataType in BlogPost Entity
- Changed the data type of CategoryId from short to int in BlogPost entity.
- Generated corresponding migration to reflect this change in the database schema.

+++ BlazorBlog.Data/Entities/BlogPost.cs
@@ -21,7 +21,7 @@
         [Required]
         public string Content { get; set; }
 
-        public short CategoryId { get; set; }
+        public int CategoryId { get; set; }
 
         public string UserId { get; set; }
